Allow tailwind config per theme
